+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 =========
 
 * :bug:`-` If ``version=False`` the console logger level was being set to zero. Now it's set to ``WARNING`` unless ``verbose=True`` which sets it to ``DEBUG`` or if ``verbose=<int>`` in which case it sets it to that numerical value.
+* :feature:`54` Filter out issues parsing out Tron replies and log them only to the file logger.
 
 * :release:`0.6.1 <2021-02-13>`
 * :feature:`-` Use log rollover.